Closes gravitee-io/issues#2406
A new configuration option 'Content-Type' was added.
This new option allows the user to select the HTTP Content-Type header value
used to invoke the backend soap service.

The following options are available:
- text/xml
- application/soap-xml
- application/xml

Before this change the policy always used 'text/xml' as the HTTP
Content-Type header value. In order to maintain backwards compatibility
the 'text/xml' was set as the default value.